refac: relocate base entity to common domain
