+++ src/main/java/org/example/kkumdoriland/KkumdoriLandApplication.java
@@ -2,8 +2,10 @@ package org.example.kkumdoriland;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 
 @SpringBootApplication
+@EnableWebSecurity
 public class KkumdoriLandApplication {
 
     public static void main(String[] args) {
