+++ src/main/java/org/example/kkumdoriland/notification/domain/Notification.java
@@ -7,7 +7,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import org.example.kkumdoriland.common.BaseEntity;
+import org.example.kkumdoriland.common.domain.BaseEntity;
 import org.example.kkumdoriland.user.domain.User;
 
 @Entity
