+++ src/main/java/org/example/kkumdoriland/request/domain/Request.java
@@ -1,6 +1,5 @@
 package org.example.kkumdoriland.request.domain;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
@@ -9,7 +8,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.EnumType;
-import org.example.kkumdoriland.common.BaseEntity;
+import org.example.kkumdoriland.common.domain.BaseEntity;
 import org.example.kkumdoriland.user.domain.User;
 
 @Entity
